requirements.txt
streamlit
pandas
numpy
python-dotenv
groq